+++ src/components/mod.rs
@@ -8,7 +8,7 @@ mod i18n_font;
 mod i18n_number;
 mod i18n_text;
 mod i18n_text_2d;
-mod utils;
+pub(crate) mod utils;
 
 pub use i18n_font::*;
 #[cfg(feature = "numbers")]