+++ src/lib.rs
@@ -3,6 +3,8 @@ mod components;
 mod interpolate;
 mod parse;
 mod plugin;
+#[cfg(feature = "plurals")]
+mod plural;
 mod resources;
 
 pub mod prelude {